Move fetchAllProjects to App
